Finished hand analaysis methods!
I added test methods for pair, two pair, three of a kind, full house, and 4 of a kind.

I also added tests for all of these methods. I also reformatted a lot of the code in Test.py (to make it easier to read).

So far, everything seems to work (all tests pass).

In my next commit, I plan on working to finish the play method.