Added code to find winner
I did this using the sorting function and a crafty use of tuples. See Poker.py:170-200.

While the code works, it's not pretty. I would like to tweak it/improve it here and there. In particular, I think the code would be better if it had a player class (and maybe a hand class?). I'll think about making changes if I have time.

Otherwise, my code does currently work, so that's good (y).